refactor: delete profiler_anonymizer
